Same as CUNY-CL/yoyodyne#347 but for UDTube.
One special thing is that we don't need anything other than the
tokenlist for prediction so we have a separate class for that, whereas
for testing we also want to be iterable but we do need the tags, so we
have a separate class for that too. That's three datasets in all, with
different features but some shared code.